fix(client): always disconnect wl_diplay
On protocol error or other connection loss the wl_display still must be
disconnected.

And this must be done only after all other objects have been destroyed
accordingly. Make sure this is possible by not freeing the wl_display
object directly.

Currently the consumer is responsible for shutting down correctly
afterwards through a correct sequence of operations, but in the future it
should be made possible to do this in the lib through a connection manager
class.