refactor(client): revise destroy logic
The logic for client shutdown was on a conceptual level wrong.

Clients should always release all their Wayland objects before closing the
connection independent of the server state.

In a similar way there is no sensible way of reconnecting after a connection
loss. It does not make sense to watch the socket file via QFileSystemWatcher.
Instead the connection teardown is communicated through the socket itself.

In a first step replace all destroy or reconnect logic in ConnectionThread and
all destroy logic in tests.

Closes #4

BREAKING CHANGE: API break of client library.