fix(server): unset focused surface on seat destruct
When the seat goes down all pointers, keyboards and so on created through the
seat become automatically defunct. But this holds not for surfaces. What
happened in autotests for example was that a defunct pointer tried to access
the seat global because the connection with the surface was not yet broken.

To circumvent this issue do break the connections between devices and surfaces
explicitly on seat destruct.