refactor(client): remove destroy method
This is a follow-up patch on an earlier patch to revise the destroy logic.
As previously noted a client has to destroy resources by calling into the
Wayland library functions independently of the compositor state.

Therefore providing a destroy method to explicitly not do this was a broken
API concept in general.

More information can be found in issue #4.

This patch removes all destroy methods and replaces them with calls to proper
release.

BREAKING CHANGE: API break of client library.