test: increase robustness of plasma window model test
The test was sporadically failing on CI. It shows that we ignore an additional
set event sent at resource creation. While this is kind of broken API do not
try to change the API for this singular issue (and risk broken legacy clients)
but instead adapt the autotest.

At some point there should be a complete new implementation on a new protocol
similar to the current one but with lessons learned in regards to Wayland
protocol design over the last few years.