ci: setup GitLab CI
Adds a GitLab CI configuration file to run checks on branches and merge
requests. This includes a linter for the commit messages, final build of the
branch and autotests.

In the future additional linters should be introduced, for example cppcheck or
a clang-format based one.

The 'ci' type is added to the list of allowed commit types.

Removes some old arc configuration files.