test: convert to guiless window management test
The window management test is a QTEST_MAIN test because it uses QPixmap at one
point. That might be a small change from a QTEST_GUILESS_MAIN test, but the
issue is that then QtWayland is fired up what opens possibilities for memory
corruptions and errors not in the scope of Wrapland.

We should aim for running all tests only as QTEST_GUILESS_MAIN. For now just
change the test to that and disable the particular QPixmap related test
portion.

Long-term the QPixmap-dependent QIcon API in Wrapland should be deprecated and
a QImage should be sent over the wire instead of a QIcon.