fix(server): ignore SIGPIPE
The design of the plasma_window_management protocol is problematic in that it
expects the client to open some unspecified file descriptor that then is used
by the server to write the icon into. This can lead to SIGPIPE.

For now just ignore the signal on a library level. But that must be done in a
library like Wrapland in a different way long-term.

Related: #7