test: disable icon tests
The concept we have for icons in the plasma window management protocol is
deeply flawed. We need to replace it better sooner than later.

But in the meantime because of the flaws tests accessing the icons by data
handle are extremely racy. Especially on CI they regularly fail.

Disable the icon-data related tests for now and work on a reconceptualization
at some point in the future.

Related: #7