fix(client): disconnect previous wl_display
A connection thread can in theory be reused and setup with a new wl_display.
While this is certainly not recommendable and our API should better forbit
it at the moment it is possible.

This change makes sure a potential earlier connection in the thread is
disconnected so that no memory is leaked.