fix(server): emit output destroy signal early
Normally the destroy signal of a global is on deletion emitted from the
global's destructor. But at this point the child class destructor is already
finished and accessing the object is a violation.

That this design is in general problematic is clear. But for now at least fix
one instance this leads to unsanitized values as the surface autotest shows:

The surface interface depends on receiving the output interface destroy signal
early so that all surfaces can emit a leave event.