test: wait on drag data offer
In the datadevice test we leak the data offer that is sent to the same surface
the drag originates from if we close the test directly without waiting on the
data offer.

It seems like a problematic design in general but must maybe solved on
libwayland level. For now let's just wait.