test: check focused pointer surface on unbound enter
We can check  at the end of the test if the focused surface has been reset
accordingly before exiting it.

In this point also a very rare data leak could be observed. It is likely that
this data leak was due to the exit before any events anymore could be
processed. If this really is the case needs to be studied more in the future.