test: set registry event queue
The registry in the sub-compositor test was created without setting the event
queue of the connection.

This lead to very rare hard to debug use after free races with the event queue
on quick destroy.

The reason for that seems to be that the connection thread dispatches events
not only for the associated event queue but also for the default one. A more
thorough investigation needs to fix the connection thread in this regard.