refactor: prototype server objects remodel project
We start the server objects remodelling effort by moving the Display,
ClientConnection and OutputInterface functionality over to the new structure.

The legacy classes are kept in place and integration paths are established to
the new classes to have some ongoing autotest capabilities of core capabilities
while other classes are moved over one by one.

The OutputInterface and Server Display tests are replaced by adapted versions
for the remodelled classes.

Continued work will be done on the server-objects branch until it is ready to
be merged back to master.

Closes #15.