docs: harmonize version numbers
All earlier Frameworks version numbers 5.XX are rewritten to fit the SemVer
approach for libraries without API/ABI guarantee and still give room for
future version numbers in the pattern of Plasma release versions (adapted to
SemVer).

For example Frameworks version 5.49 is replaced with 0.0.549 and Plasma
version 5.18 is replaced with 0.518.0.

Closes #2.