ci: block detached pipelines from merge requests
Counter-intuitively when there are jobs with rules a detached merge requeust
pipeline is created including every job that has rules.

It is necessary to block detached pipelines from merge requests to fight this
behavior. [1]

This commit is analogous to a commit in KWinFT where the behavior was
experienced with a merge request. It is assumed the same problem would hit us
here in Wrapland at some point too.

[1] https://docs.gitlab.com/ee/ci/yaml/#exclude-jobs-with-rules-from-certain-pipelines