revert: test: for remote access params try-compare on second spy
This reverts commit 1fd362051f6a98fe818a9e21c966e65bf7914d02.

The try-compare does not seem to work. We need to go back to the
event loop when the second signal has not yet been received. But the
solution here might also fail when the other thread writes to the signal
between our conditional check and the signal wait.

Therefore this is not a proper fix but makes it just less likely. A proper
fix would be using mutex and atomic values when checking against the spies.
Something that needs more work.