ci: fetch all from repo for message lint
When the pipeline runs on tags it can not find the master head. and then fails
because commitlint tries to lint all commits.

This is likely because we only fetch the HEAD and not the whole history. So try
to fetch everything instead. Since we cache the repo it should be resonable
fast later on.