ci: add script for stable release
Adds a script for the final release version from beta version. It uses the
previous tag on master, that should be the latest beta tag. Also adds the
changelog update back to master branch.

As with the beta release the process is manually at the moment, is started
from master branch and needs a final git push --follow-tags to the stable
branch.