test: disable check on second remote access param
The check failed sporadically but more than often enough on CI under high load.
The spy connection is just too racy. That holds for all signal spies but here
it really happens too often.

Long-term we need to find a general solution for this problem. That is also a
prerequisite for solving issue #12.