ci: build only after lint
We can do lint and build in parallel but if the lint fails the build will just
keeps going. Assuming there will be many issues in the beginning with commit
messages it is wasteful to do it this way.

The lint does not need very long so just do it in sequence such that the build
won't start if the lint fails.