ci: lint not on upstream branches
It is problematic to lint commits on upstream branches in case a commit with
wrong format was pushed. Or when there is some bug in the linter, what is
currently the case, see [1][2].

Make sure the linter is only always run on merge requests and for other
branches.

[1] https://github.com/conventional-changelog/commitlint/issues/896
[2] https://gitlab.com/kwinft/wrapland/-/jobs/495110391

(cherry picked from commit 99803545c72eb988257bfa685aab185c3093a1f0)