refactor: remodel input classes
This patch remodels KeyboardInterface and PointerInterface with its dependent
classes pointer constraints, pointer gestures and relative pointer in the new
server object model.

A new signal is introduced to communicate when a resource is destroyed.
Currently this must be added to every public API class. A future step could be
to move it in a common parent class.