refactor(server): remodel surface and related classes
This moves Server::Surface (before SurfaceInterface) to the new structure.

Related classes like Compositor, Subcompositor, Subsurface and Region are
moved too. This was necessary for compilation to succeed.

Many more autotests are failing now but since the change is already very big,
let that be and deal with it afterwards.

Closes #27.