refactor(server): remodel buffer and dmabuf classes
Move over these classes to the new server model. In particular the Buffer
class is a central one that needs to be moved now as Surface was moved.

A new class BufferManager is introduced for managing the wl_shm access per
Display.

The dmabuf implementation has been streamlined but will need more work in the
future. In particular it needs an autotest.

Closes #40.