fix(server): for Surfaces check for null region and leave Outputs correctly
An input or opaque region might be null. That must be checked to not seg fault.

Also the outputs vectors were not cleaned up correctly with erase. Do this now
and also connect to the new removed signal to send leave signals before the
Output's wl_global goes down.

This fixes the Surface autotest.