fix(server): on Surface resource destroy unfocus pointer
We connect to Surface's resourceDestroyed signal and clean up the Seat
implementation here too.

The Seat autotest is fixed in that one more metatype must be registered and
we do not try to enter an unbound surface. Compositors must respect this and
listen for the resoureDestroyed signal accordingly we say for now.