refactor(server): remodel xdg-shell
The deprecated protocols wl_shell and xdg-shell v6 are removed. The stable
protocol xdg-shell is moved to the new server objects model.

Here we remove all the abstraction that had been introduced in the past
(GenericShellSurface, SurfaceRole) and bring it back on the most basic level
of one class per wl_global/wl_resource. We can think about introducing some
abstraction back again later on.

As xdg-decoration depends on xdg-shell it is directly remodeled here too.

Closes #45.