fix(server): omit flushing the client on drop
When the client goes away the client does not exist anymore while a drop on the
resource is still possible.

Instead of checking if the client is still available just omit flushing it for
now and see if this was necessary at all. The assumption is that it is not.