ci: update clang-tidy configuration
This moves the clang-tidy configuration into a separate file in the source
root directory. The script must be called from this directory and the path to
the build directory provided, otherwise it is assumed the path is `../build`.

The clang-tidy checks we want to run are selected in the configuration file.