fix(server): return when adding socket fails
In order to not directly crash the compositor process abort when the socket is
not created with success.

Internally we do this by throwing an exception. We could rethrow this assuming
the consumer will handle it.

For now we just print a warning though to the log and abort. That assumes the
consumer checks afterwards running() as it was in the past.

The noXdgRuntimeDir test passes again.