refactor(server): define special member functions
In Wayland namespace classes with dtor define the other special member
functions according to the rule of five.

For capsules define the move operations for others for now delete all. We will
later see if this holds up.

Since the Display class is an abstract class it is a special case and we will
handle it later.