test: ensure unique timestamps
The seat test was randomly failing on high load because we compared the current
time on two calls to the system one after the other what can lead to comparing
different time values.

To make sure we also have different values in comparision to an earlier button
press event just use the previous time value incremented.