docs(server): annotate TODOs in code with author information
Every TODO entry needs author information such that the person responsible can
be identified easily.

For old TOODs where we did not have this rule yet and no author was easily
identifiable we set "unknown author" but in TODOs we will require the author
to always add their name.

This fixes the google-readability-todo clang-tidy check.