style(server): use variadic arguments for Qt logging
Instead of stream operators use the variadic arguments that Qt logging macros
allow. The reason for doing it this was is that the macros are otherwise not
C++ standard compliant.

This fixes the clang-diagnostic-gnu-zero-variadic-macro-arguments clang-tidy
check. Also a compiler warning might be generated by its usage.

For few selected usages where stream operators are more easy to use the
clang-tidy check has been disabled.

Long-term we want to move away from making use of Qt logging either way.