build(server): ignore selectively clang-tidy alignment checks
In two cases in our code a cast results in a memory alignment change according
to clang-tidy.

In the wl_container_of case we probably can't do much about it so just disable
the clang-tidy check there.

In the EglStreamController class we load a struct from a library dynamically.
There might be a solution to that but it is not directly apparent. Since it is
legacy code just disable the check for now in this case too.