style(server): name parameters the same
In declration and definition of a function the parameters should be named the
same.

The clang-tidy check readability-inconsistent-declaration-parameter-name passes
now but in Output and OutputDeviceV1 classes the linter is deactivated on some
functions that are defined through macros and by that have different parameter
names.

Long-term we want to get rid of any such macro-usage and replace it anyway with
templates so for now we just ignore the lint in these specific cases.