refactor(server): replace or decorate static-capable functions
The clang-tidy check readability-convert-member-functions-to-static reported
several functions as being unnecessarily a non-static member function.

In most cases this meant that the function could be a file-scoped helper
function or was superfluous in the first place.

In one case the static keyword was added but a TODO comment was added to mark
that this function might be a candidate for future refactoring.

The aforementioned clang-tidy check passes now.