refactor(server): remove static list of clients
The Wayland::Client objects were referenced in a static vector. This was a
suboptimal design for a library. This static container is removed. On destroy
we reference the client implicitly by a wrapper container that we can
extract in the normal libwayland way of calling the macro wl_container_of.