ci: filter and enforce clang-tidy checks
Filter out some clang-tidy checks that are at the moment or in general
unsuitable to run on the server library. For every disabled check an
explanation is provided in the top-level clang-tidy configuration file.

Enforce passing of all other checks in the CI. The pipeline will fail now if
one or more checks are not passing.