docs: use modern style MIT license text with correct authors
The output device and management protocol xml files used an unusual and old
style of the MIT license text. Switch to the common modern style as described
in [1][2].

Looking back at the first commit bcf03a1b the original author of the
predecessor protocol likely made a mistake in naming other authors besides
himself. Because of that and since the KWinFT protocol versions are only loosly
based on the predecessor protocol this commit therefore also correct the list
of copyright holders.

[1] https://opensource.org/licenses/MIT
[2] https://fedoraproject.org/wiki/Licensing:MIT