feat: provide method for server-side resource destruction
Some protocol objects might be destroyed automatically by the server if a certain
condition described in the protocol is fulfilled. For such objects introduce a new method
of resource destruction. The destructor is set to null in this case so the server-side
Resource wrapper object and the handle must be destroyed afterwards explicitly.