test: wait for selection signal to not leak memory
The selection test was sporadically leaking memory in the client library due to not
waiting for the selection signal. In general this should not leak memory but it is
unclear what the root cause is. For now add a wait on the selection signal to have
a reliable test behavior.