fix(server): release only surface buffers automatically
A compositor might create wl_buffers internally and then request a Buffer handle for
these. In this case we do not want to automatically release the wl_buffer on Buffer
handle destroy but let the API consumer decide when it is time. It can still enable
automatic release by calling setCommitted on the Buffer.