feat(server): add master output class
The previous several classes for wl_output, xdg-output and output_device_v1 in
the Server library that a compositor had to interact with are replaced with a
single master class that mediates and controls functionality of the other
classes.

A compositor only needs to deal with this class instead of all the other ones.
Additionally the update procedure is improved by that: the done() call has a
centralized entrance point and internally the Output class forwards this to
other classes if they are enabled. Changes are applied atomically throughout.

BREAKING CHANGE: The Server Output class is refactored as a single access
point for the compositor.