feat: remove remote-access
Remote-access was an early attempt at creating a Wayland protocol for screen
casting. KDE's xdg-desktop-portal backend has now switched to another approach
with Pipewire internal to KWin.

Since the remote-access protocol never gained traction outside of the
xdg-desktop-portal backend it becomes without real users and since it was used
in the past rather sparsely and by that is not tested very well it makes no
sense to continue supporting it in Wrapland.

BREAKING CHANGE: Legacy remote-access protocol and API are removed