refactor(server): modernize subsurface callback
The subsurface creation callback used some outdated internal API. We can
improve it by putting everything into the static callback and acting on the
bind directly.

This allows us to remove a function marked "legacy" in the Global class that
was only used still by the outdated subsurface callback code.