fix(server): use global bind versions on resource creation
When a resource is created through a global interface the version of the
resource is determined by the version of the bind which might be lower than
the version advertised by the compositor depending on what the client supports.

For that reason the version provided to the resource must always be the one of
the bind. We change this up for current implementations that did it wrong and
remove the version getter from the Global class what ensure it is not used
anywhere else.