fix(server): delay global destroy
A known issue of libwayland is that the destruction of globals is racy as
clients might still connect to them via the registry while the compositor has
already destroyed them.

To mitigate this issue with [1] new library API was introduced to remove a
global from clients while still keeping it alive on the compositor for an
unspecified time.

We do this now and wait for 5 seconds before destroying the object for real
what should give any client enough time to cease communication in relation to
the removed global.

[1] https://gitlab.freedesktop.org/wayland/wayland/-/merge_requests/28