fix(server): check for global being removed and destroyed
On every request to a global we need to check if it has been removed before.
We do this via a small templated wrapper that can be easily integrated to
existing classes. We do this first for the blur and contrast managers since
there was an autotest using these types to check the failing condition in the
past and will now pass.