feat(client): simplify current mode logic
Instead of using flags to indicate preferred and current modes use:
* a single boolean for the preferred propertey
* the current-mode getter to define what the mode is.

The list of modes stays in the same order over the whole runtime. A change of
the current mode does not change this order but only sets a different current
mdoe internally that then can be retrieved by the current-mode getter.

This also presumably fixes an issue with modes "vanishing" after being
switched. It is assumed the modes were lost internally due to the reordering.

BREAKING CHANGE: The client output device mode API changes.