feat: streamline output device information
The output device information is streamlined to send the following information:
- name, same as in xdg-output protocol.
- description, same as in xdg-output protocol.
- make, the manufacturer.
- model, the display model if available.
- serial_number, the serial number of the display.

The arguments name and description mirror the xdg-output v2 protocol. Here
description is supposed to be pre-translated by the compositor. The arguments
make, modle and serial_number allow clients to identify a display over multiple
sessions and load back a previous configuration.