feat: support xdg-output v3
Add support for version 3 of the xdg-output interfaces. No new events or
requests are added, instead the transmition logic is changed in that
wl_output::done is now sent instead of the done event of xdg-output which is
deprecated. Additionally the description field becomes mutable.

For the change in transmition we need to add some internal logic to the server
class and connect done in the client library to the wl_output::done induced
change signal of Output.