fix(client): clean up wlr output modes with smart pointer
The modes objects leak if we do not clean them up when the output is removed.
Do this with unique_ptrs. For that we switch from a QVector to std::vector
to store them.

(cherry picked from commit 9786f0f0d4c92f5a22d56633d58c08ad60a226cf)