feat(server): provide effective window geometry from xdg-shell surface
Compositors should only be interested in the effective window geometry of an
xdg-shell surface. That is the geometry pre-clamped to the full expanse of the
wl_surface with all its subsurfaces.

For that we provide a new helper function in the Surface class to get the full
expanse and move the window geometry getter for simplification from the
toplevel and popup classes to the shell surface class.

BREAKING CHANGE: Server library's xdg-shell surface class provides effective
window geometry.