build: move release scripts to tooling directory
With the latest changes to our CI infrastructure the old ci directory has
become mostly empty.

And the release scripts are better described as tooling anyways because they
have to be executed locally. Move them and remaining links over to the tooling
directory such that the ci directory is emptied out.