fix(server): omit sending keyboard leave on client destroy
When the client of the currently focused surface has been destroyed do not try
to send a leave event referencing this client.

This can happen as libwayland sends the client destroy signal before destroying
all associated resources. Long-term another option could be upstream to provide
a second signal that is emitted after all resources have beend destroyed and
then listen for that. That would need an upstream change in libwayland. For now
this patch fixes the issue when a client has keyboard focus at the moment it
gets destroyed.