style(client): rename XdgShellSurface to XdgShellToplevel
In stable xdg-shell the shell surface is not directly associated with the
toplevel and we should therefore not name the class like this. It is after all
an xdg-shell toplevel only.

The factory function is renamed as well. We use snake-case as we plan to do
with at least function names in the future in general.