fix(server): allow getting xdg-popup without parent
Per protocol it is allowed that a client creates an xdg-popup without directly
specifying a parent. The idea behind this is that the client may set a parent
later on through other means.

To allow this only return with error when a parent was specified but no
internal representation was found for it. If the parent was null, continue.