feat(server): add support for the drm_lease_v1 protocol
The protocol provides a communication channel for clients to request leases on
DRM resources.

We provide a public interface with a DRM lease device global, lease connectors
and finally client-side created leases.

The compositor needs to create a lease connector for every output it might
allow to lease to clients. In the end the compositor decides that though on a
per-lease basis by issuing grant or rather finish on any lease, that has been
created by a client.