fix(server): hold onto appmenus in order to send them to new resources in PlasmaWindow
The setter function was const, which meant that it
did not mutate m_applicationMenu in order to update
it to the new values.

In createResource, this means that it will never send
the appmenu to new resources, because m_applicationMenu
was always blank due to the setter not changing it
when it should.

BREAKING CHANGE: PlasmaWindow API changes