feat(server): remove capabilities signals
The pointer, keyboard and touch capabilities are set by the consumer. It does
not make sense to emit a signal on change to inform the consumer about the
change.

In light of brevity remove the signals and call the sendCapabilities function
directly instead.

BREAKING CHANGE: seat capabilities signals are removed