refactor(server): make input device pools optional
According to the Wayland specs wl_pointer, wl_keyboard and wl_touch created
before the latest respective capability switch on should be ignored.

We can map that by providing their pools as std::optionals in the Seat class.
For now the value is still always set in the constructor but next the value
should be dynamically set on the call to the respective capability setter.