refactor(server): restrict access to position variable
Since the pointer_pool is now in the public interface, restrict access to the
pos variable. It should only be set via a setter function that then also emits
a signal.

We already had a getter for the position which we rename to fit the
set_position name and use outside the class now always.