refactor(server): restrict access to keyboard devices variable
Since the keyboard_pool is now in the public interface, restrict access to the
devices variable. It should only be set from within the class on device
creation. We do not access the variable outside of the pool, so we do not have
to add a getter function for it.