refactor(server): remove subsurface position workaround
The comment says that it is necessary to set the position of the subsurface
immediately because of a Qt bug. This Qt bug got closed some years ago.

Removing the direct position update requires us to be more thoughtful on parent
commits. Even if we are not doing synced updates on the subsurface we must
update the position.

Also there was an issue with the set position call. We compared against the
wrong position variable. Otherwise the positions of desync subsurfaces are not
guaranteed to be updated. For example Firefox on fullscreen moving a
subsurface back to (0,0) failed. Correct this and compare against the current
position, not the scheduled one.