style(server): rename focus surface destroy connections
The connections are renamed from destroy_connection to surface_lost_notifier to
make more clear what it is about.

Since it's an implementation detail its definition is moved to the end of the
structs. At some point in the future the connections should go into a separate
struct that is only declared in the public interface but defined in a details
namespace.