fix(server): cancel old selection source after setting new one
When we set a new selection the old selection must be cancelled. Do this after
the new selection has already been announced. This ensures that clients
listening in on selection changes get informed about the new selection and do
not receive an event informing them only about the destruction of the old one.

This is important as a client might react to such a destruction event by
resubmitting the content (clipboard managers keeping the seleciton content up
after the source had been destroyed).