feat(server): replace proxy remote surfaces with drag movement block
With the new external data sources the proxy remote surfaces mechanism becomes
useless. Instead we replace it with a simple variable on the drag to indicate
if the source surfaces still expect pointer movement events to reach them. The
default is that such events are filtered out. Consumers must change that after
start of a drag with a call to drag_pool::set_source_client_movement_blocked.

BREAKING CHANGE: proxy remote surfaces removed