feat(server): set keymap as raw C string
It is better to provide a simple interface for handing over a char pointer for
the keymap instead of wrapping it into a std::string. That minimizes copies and
since compositors usually receive the keymap as a char pointer from xkbcommons
it is similarly comfortable for consumers as an interface with std::string.

We can also check this way easily if the set keymap is really a new one.
Otherwise we can omit resending the keymap.

BREAKING CHANGE: keymaps are set as C strings