perf(server): lazy-send keymaps on focus change
When the a new wl_keyboard is created and when the keymap changes we don't have
to directly submit to the new device or all existing devices. The keymap is
only relevant when receiving keyboard input, so on keyboard focus.
Therefore wait with sending keymap events to keyboards until the keyboard gets
focus. Especially on keymap changes it is expected this means to create way
less string copies of the keymap to send to all keyboards with focus only.