feat(server): store virtual desktops in STL vector
Replace the QList with an std::vector. We simplify the implementation by
removing one of the find desktop functions and making the remaining one a free
function in the source file only. Sending the removed signal is put into a
separate member function in the Private class to access it from the destructor
directly instead of going through the interface function.

BREAKING CHANGE: virtual desktops getter returns STL vector instead of QList