feat(server): allow to create DRM lease connector without an output
Compositors usually want to advertise connectors to DRM lease clients that are
not at the same time associated with an wl_output. For such connectors provide
a separate connector factory in the DRM lease device class that only takes the
basic properties of the connector instead of an Output type object.

Internally the basic data is now saved in both cases. If a DRM lease connector
was created from such basic data or an Output can be checked by getting the
output pointer from it.