fix(server): assume input devices are created and ensure they are removed
Checking if the call to a new input device returns a non-null object does not
make sense. Instead we should throw and catch an exception if there is a
problem on creating an input device. This is to be done in the future.

On the other side though we can check that removing only one input device is
enough since we only added one. Ensure this through an assert.