fix(server): accept buffer damage unconditionally
We used to check on the surface having a buffer update but it is not specified
if the buffer must first be attached before sending buffer damage or it can
also be done the other way around. Firefox does this sometimes and to register
the damage correctly we must set it if the buffer has not yet been updated too.