feat: update to Plasma window management version 16
The new version adds the `resource_name_changed` event for XWayland windows.

Client-side, we don't store the new name. We emit an event with std::string payload,
the client is welcome to grab it.
Server-side, the function takes a const reference to string. It is safe to assume this
string is an attribute on the window.

Note:
Only the org_kde_plasma_window interface bumps version. It is now ahead of the global
creating it.