feat(server): set one modifier per dmabuf buffer
The modifier mechanism in the kernel had been designed initially to allow one
modifier per plane. That's why the modifier of a params object is sent via the
add request. But this was changed later on to only allow one modifier per
buffer [1].

Change our public API and internal code and do it like wlroots does it [2]:
store the modifier only once and send an invalid format error back in case the
client sends different modifiers on the same params object.

[1] https://github.com/torvalds/linux/commit/bae781b259
[2] https://gitlab.freedesktop.org/wlroots/wlroots/-/commit/28020ff5

BREAKING CHANGE: dmabuf modifier set once per buffer, not for each plane