style: enforce east const
Clang-Format 14 comes with the new option "QualifierAlignment" which allwos to
enforce west- or east-const.

Despite the C++ Core Guidelines still recommending to use west const because it
is more common [1] we moved to using east const in the past whenever such a
statement had been modified anyway or a new one was added. Because there are
good arguments for using east const consistently. [2] [3]

The new clang-format option gives us the chance to use east const consistently
and enforce it in the future. Do this now by switching over all remaining west
const qualifiers via clang-format inline edits and enforce it on CI.

[1] https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rl-const
[2] https://mariusbancila.ro/blog/2018/11/23/join-the-east-const-revolution/
[3] https://www.youtube.com/watch?v=fv--IKZFVO8