feat: support org_kde_plasma_activation_feedback interface
The protocol is used to show feedback on activation in a Plasma desktop. When
an app is about to be launched but no surface yet exists, the desktop will show
a loading indicator or similar UI elements instead.

The app id sent in this protocol corresponds to the application id in the
XDG activation protocol.