refactor(server): publish KDE idle timeout objects
Instead of handling the idle objects internally export them to consumers.
Consumers are from now on expected to handle these objects according to their own
policies and to decide on respecting inhibition requests or not.

This gives consumers more freedom in implementing the mechanism and reduces the
size of our internal implementation what is in line with the general scope of
Wrapland being a wrapper library around libwayland.

BREAKING CHANGE: KDE idle API changes