build: set default build type
The CMAKE_BUILD_TYPE variable is set to an empty string by default. [1]
But KDECompilerSettings did set it to a reasonable value before 13f5ea9f. As
some tooling won't work correctly without it fix this regression by using a
minimal template for a BuildType.cmake module which does the same. [2]

[1] https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_TYPE.html
[2] https://www.kitware.com/cmake-and-the-default-build-type/